LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/blob.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://www1.web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://www1.web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to blob:https://www1.web-platform.test:9443/c279a356-5184-404c-867d-a7f50d6d4a88 because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cancelled load to blob:https://not-web-platform.test:9443/0c0cce0b-c04e-4fc0-b76d-8385d871521c because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Cancelled load to https://www1.web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://not-web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://not-web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://web-platform.test:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/common/blank.html due to access control checks.

Harness Error (TIMEOUT), message = null

PASS Cross-Origin-Embedder-Policy and blob: URL from https://web-platform.test:9443 in subframe via subframe
PASS Cross-Origin-Embedder-Policy and blob: URL from https://web-platform.test:9443 in subframe via navigate
PASS Cross-Origin-Embedder-Policy and blob: URL from https://web-platform.test:9443 in subframe via popup
PASS Cross-Origin-Embedder-Policy and blob: URL from https://www1.web-platform.test:9443 in subframe via subframe
TIMEOUT Cross-Origin-Embedder-Policy and blob: URL from https://www1.web-platform.test:9443 in subframe via navigate Test timed out
PASS Cross-Origin-Embedder-Policy and blob: URL from https://www1.web-platform.test:9443 in subframe via popup
PASS Cross-Origin-Embedder-Policy and blob: URL from https://not-web-platform.test:9443 in subframe via subframe
TIMEOUT Cross-Origin-Embedder-Policy and blob: URL from https://not-web-platform.test:9443 in subframe via navigate Test timed out
PASS Cross-Origin-Embedder-Policy and blob: URL from https://not-web-platform.test:9443 in subframe via popup
